+++ Rdmp.Core/CommandExecution/AtomicCommands/ExecuteCommandAddDatasetsToConfiguration.cs
@@ -62,7 +62,7 @@ public class ExecuteCommandAddDatasetsToConfiguration : BasicCommandExecution
                 var _importableDataSets = childProvider.ExtractableDataSets.Except(_datasets)
 
                     //where it can be used in any Project OR this project only
-                    .Where(ds => !ds.Projects.Any()|| ds.Projects.Select(p => p.ID).Contains(targetExtractionConfiguration.Project_ID))
+                    .Where(ds => (!ds.Projects.Any() || ds.Projects.Select(p => p.ID).Contains(targetExtractionConfiguration.Project_ID)) && !ds.Catalogue.IsInternalDataset)
                     .ToArray();
 
                 SetExtractableDataSets(true, _importableDataSets);
@@ -97,8 +97,10 @@ public class ExecuteCommandAddDatasetsToConfiguration : BasicCommandExecution
             }, _toadd.Cast<ExtractableDataSet>().ToArray(), out var selected))
                 return;
 
-            foreach (var ds in selected)
+            foreach (var ds in selected.Where(ds => BasicActivator.IsInteractive && ds.Catalogue.IsDeprecated && YesNo($"{ds.Catalogue.Name} is deprecated. Are you sure you wish to extract it?", "Confirm use of Deprecated Catalogue")))
+            {
                 _targetExtractionConfiguration.AddDatasetToConfiguration(ds);
+            }
         }
         else
         {