+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [9.0.2] - Unreleased
 - Add ability to use archive trigger when re-releasing to a database
+- Add ability to view logs as a flat list. Default settings can be updated via user settings
+- allow the use of Catalogue CIC filters in coresponding Extractions
+- Add Data Load component to allow SQL to be executed on an external database server
+- Improve Extraction Log Viewer Filter
 
 ## [9.0.1] - 2025-07-31
 - Improve Right-click performance