+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [9.0.2] - Unreleased
+## [9.0.2] - 2025-10-13
 - Add ability to view logs as a flat list. Default settings can be updated via user settings
 - Add Data Load component to allow SQL to be executed on an external database server
 - Improve Extraction Log Viewer Filter